Did some refactoring of the bot code so that it's inclusion into either the websocket or flask version is a lot easier and there is only one place to edit.
