+++ Dockerfile.flask
@@ -11,7 +11,7 @@ WORKDIR /app/
 COPY --from=builder /usr/local/bin/ /usr/local/bin/
 COPY --from=builder /usr/local/lib/ /usr/local/lib/
 COPY *.py /app/
-ENTRYPOINT gunicorn --bind :$PORT --workers 1 --threads 2 --timeout 0 main:flask_app
+ENTRYPOINT gunicorn --bind :$PORT --workers 1 --threads 2 --timeout 0 main_flask:flask_app
 
 #
 # docker run -e SLACK_SIGNING_SECRET=$SLACK_SIGNING_SECRET -e SLACK_BOT_TOKEN=$SLACK_BOT_TOKEN -e PORT=3000 -p 3000:3000 -it your-repo/hello-bolt
