+++ Dockerfile.websockets
@@ -4,4 +4,5 @@ COPY requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
 COPY . .
 EXPOSE 8080
-CMD ["python", "slack_gpt_bot.py"]
+# CMD ["python", "slack_gpt_bot.py"]
+CMD ["python", "main_websocket.py"]
