+++ main_websocket.py
@@ -1,25 +1,24 @@
+import openai
+import os
 import logging
 from json_logger_stdout import json_std_logger
+# from dotenv import load_dotenv
+from slack_bolt import App
+from slack_bolt.adapter.socket_mode import SocketModeHandler
 from collections import namedtuple
 from functools import partial
 
+
 from utils import (N_CHUNKS_TO_CONCAT_BEFORE_UPDATING, OPENAI_API_KEY,
+                   SLACK_APP_TOKEN, SLACK_BOT_TOKEN,
                    num_tokens_from_messages, process_conversation_history,
                    update_chat)
 
-
-from slack_bolt import App
-import openai
-
-#DEBUG Configuration
-logging.basicConfig(level=logging.INFO)
-json_std_logger.setLevel (logging.INFO)
-
-app = App()
+app = App(token=SLACK_BOT_TOKEN)
 openai.api_key = OPENAI_API_KEY
 
 ################################################
-# GPT 3.5 Models
+
 OPENAI_MODEL_DEFAULT = "gpt-3.5-turbo"
 OPENAI_MODEL_MAX_TOKENS = 4096
 
@@ -27,26 +26,7 @@ OPENAI_MODEL_CROSSOVER_POINT = OPENAI_MODEL_MAX_TOKENS * 0.75
 
 OPENAI_MODEL_EXTENDED = "gpt-3.5-turbo-16k"
 OPENAI_MODEL_EXTENDED_MAX_TOKENS = 16384
-#-----------------------------------------------
-# GPT 4 Models
-OPENAI_MODEL_4_DEFAULT = "gpt-4"
-OPENAI_MODEL_4_MAX_TOKENS = 8192 - 1
 
-OPENAI_MODEL_4_CROSSOVER_POINT = OPENAI_MODEL_4_MAX_TOKENS * 0.75
-
-OPENAI_MODEL_4_EXTENDED = "gpt-4-32k-0613"
-OPENAI_MODEL_4_EXTENDED_MAX_TOKENS = 32768 - 1
-
-'''
-https://help.openai.com/en/articles/7102672-how-can-i-access-gpt-4
-
-8/1/23: We are not currently granting access to GPT-4-32K API at this time, 
-but it will be made available at a later date.
-'''
-OPENAI_MODEL_4_EXTENDED_FEATURE_FLAG = False
-#-----------------------------------------------
-# Model to use
-OPENAI_MODEL_IN_USE = OPENAI_MODEL_4_DEFAULT
 ################################################
 def get_conversation_history(channel_id, thread_ts):
     return app.client.conversations_replies(
@@ -92,20 +72,11 @@ $0.003 per 1K input tokens and $0.004 per 1K output tokens.
 So if the conversation exceeds the cutoff, then switch to using the extended model. Otherwise, use 
 the standard model, as that seems fine for most conversations at this point.
 '''
-def determine_openai_model_3_5_to_use(input_token_count):
+def determine_openai_model_to_use(input_token_count):
     if input_token_count > OPENAI_MODEL_CROSSOVER_POINT:
         return (OPENAI_MODEL_EXTENDED, OPENAI_MODEL_EXTENDED_MAX_TOKENS)
     else:
         return (OPENAI_MODEL_DEFAULT, OPENAI_MODEL_MAX_TOKENS)
-
-def determine_openai_model_4_to_use(input_token_count):
-    if OPENAI_MODEL_4_EXTENDED_FEATURE_FLAG:
-        if input_token_count > OPENAI_MODEL_4_CROSSOVER_POINT:
-            return (OPENAI_MODEL_4_EXTENDED, OPENAI_MODEL_4_EXTENDED_MAX_TOKENS)
-        else:
-            return (OPENAI_MODEL_4_DEFAULT, OPENAI_MODEL_4_MAX_TOKENS)
-    else:
-        return (OPENAI_MODEL_4_DEFAULT, OPENAI_MODEL_4_MAX_TOKENS)
 
 def stream_openai_response_to_slack(openai_response, slack_update_func):
     response_text = ""
@@ -129,8 +100,6 @@ def handle_app_mentions(body, context):
     max_response_tokens = None
     channel_id = None
     thread_ts = None
-    model = None
-    token_count = None
     try:
         logging_wrapper("Arguments", logging.DEBUG, body=body, context=context)
 
@@ -139,9 +108,6 @@ def handle_app_mentions(body, context):
         bot_user_id = context['bot_user_id']
         user_id = context['user_id']
 
-        logging_wrapper("Milestone", logging.DEBUG, 
-                milestone="Fetching user information from slack",
-                user_id=user_id)
         user = get_user_information(user_id)
 
         '''
@@ -160,33 +126,17 @@ def handle_app_mentions(body, context):
             )
             return
         '''
+
         slack_resp = app.client.chat_postMessage(
             channel=channel_id,
             thread_ts=thread_ts,
             text=build_personalized_wait_message(user.first_name)
         )
 
-        logging_wrapper("Milestone", logging.DEBUG, 
-                milestone="Fetching conversation history from slack",
-                email=user.email,
-                channel=channel_id,
-                thread_ts=thread_ts)
-
         reply_message_ts = slack_resp['message']['ts']
         conversation_history = get_conversation_history(channel_id, thread_ts)
-
-        logging_wrapper("Milestone", logging.DEBUG, 
-                milestone="Processing conversation history from slack",
-                email=user.email,
-                channel=channel_id,
-                thread_ts=thread_ts)
         messages = process_conversation_history(conversation_history, bot_user_id)
-
-        logging_wrapper("Milestone", logging.DEBUG, 
-                milestone="Counting tokens",
-                messages=messages)
-        # num_conversation_tokens = num_tokens_from_messages(messages, OPENAI_MODEL_DEFAULT)
-        num_conversation_tokens = num_tokens_from_messages(messages, OPENAI_MODEL_IN_USE)
+        num_conversation_tokens = num_tokens_from_messages(messages, OPENAI_MODEL_DEFAULT)
         
         '''
         print(openai.Model.list())
@@ -218,25 +168,12 @@ def handle_app_mentions(body, context):
         },
         '''
 
-        logging_wrapper("Milestone", logging.DEBUG, 
-                milestone="Determining OpenAI Model to use",
-                num_conversation_tokens=num_conversation_tokens)
-
         #Pick the model to use based on the number of tokens used thus far
         #picking the extended model, means spending more, so only select that when 
         #necessary
-        # model, token_count = determine_openai_model_3_5_to_use(num_conversation_tokens)
-        model, token_count = determine_openai_model_4_to_use(num_conversation_tokens)
+        model, token_count = determine_openai_model_to_use(num_conversation_tokens)
 
         max_response_tokens = token_count-num_conversation_tokens
-        logging_wrapper("Milestone", logging.DEBUG, 
-                        milestone="Forwarding request to OpenAI",
-                        model_used=model,
-                        token_count=token_count,
-                        token_used_count=num_conversation_tokens,
-                        email=user.email,
-                        request=messages[-1])
- 
         openai_response = openai.ChatCompletion.create(
             model=model,
             messages=messages,
@@ -261,7 +198,6 @@ def handle_app_mentions(body, context):
     
     except Exception as e:
         logging_wrapper("Exception", logging.ERROR, 
-            model_used=model,
             token_used_count=num_conversation_tokens,
             max_response_tokens=max_response_tokens,
             channel_id=channel_id, 
@@ -276,20 +212,7 @@ def handle_app_mentions(body, context):
             thread_ts=thread_ts,
             text=f"Sorry, I can't provide a response. Encountered an error:\n`\n{e}\n`")
 
-# @app.command("/hello-bolt-python")
-# def hello(body, ack):
-#     user_id = body["user_id"]
-#     ack(f"Hi <@{user_id}>!")
-
 ################################################
-from flask import Flask, request
-from slack_bolt.adapter.flask import SlackRequestHandler
-
-flask_app = Flask(__name__)
-handler = SlackRequestHandler(app)
-
-
-@flask_app.route("/slack/events", methods=["POST"])
-def slack_events():
-    print('slack event received')
-    return handler.handle(request)
+if __name__ == "__main__":
+    handler = SocketModeHandler(app, SLACK_APP_TOKEN)
+    handler.start()
